requirements.txt
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.15.2
statsmodels==0.14.1
tqdm==4.65.2
